Revert back to older version of punchHoles
The old version was way faster, and just as capable of generating puzzles at each desired difficulty.
The random nature of the design of the new search caused a lot of wasted computation despite not generally being able to find puzzles of more difficulty than the old one could.
Since the old one is about 10 times as fast (or 100, if using the high iteration count), the old one can fill the arrays much faster anyway.

Additionally, adds a small optimization in the solver.