Fix bug where easier puzzles can trickle into harder difficulties
Due to a flaw in the logic that prevents runaway storage consumption, if a difficulty already has 100 puzzles, any new puzzles of that difficulty may be misclassified as harder puzzles. For example, if there are 100 medium puzzles and a new medium puzzle is generated, it will skip past and place the puzzle into the hard puzzle list.